Create new locale directory
